+++ photutils/aperture/bounding_box.py
@@ -41,11 +41,11 @@ class BoundingBox:
     >>> bbox == BoundingBox(ixmin=7, ixmax=10, iymin=2, iymax=20)
     False
 
-    >>> # "shape" and "slices" can be useful when working with numpy arrays
+    >>> # "center" and "shape" can be useful when working with numpy arrays
+    >>> bbox.center  # numpy order: (y, x)
+    (10.5, 5.0)
     >>> bbox.shape  # numpy order: (y, x)
     (18, 9)
-    >>> bbox.slices  # numpy order: (y, x)
-    (slice(2, 20, None), slice(1, 10, None))
 
     >>> # "extent" is useful when plotting the BoundingBox with matplotlib
     >>> bbox.extent  # matplotlib order: (x, y)
@@ -113,7 +113,6 @@ class BoundingBox:
         >>> BoundingBox.from_float(xmin=1.4, xmax=10.4, ymin=1.6, ymax=10.6)
         BoundingBox(ixmin=1, ixmax=11, iymin=2, iymax=12)
         """
-
         ixmin = int(np.floor(xmin + 0.5))
         ixmax = int(np.ceil(xmax + 0.5))
         iymin = int(np.floor(ymin + 0.5))
@@ -126,12 +125,10 @@ class BoundingBox:
             raise TypeError('Can compare BoundingBox only to another '
                             'BoundingBox.')
 
-        return (
-            (self.ixmin == other.ixmin) and
-            (self.ixmax == other.ixmax) and
-            (self.iymin == other.iymin) and
-            (self.iymax == other.iymax)
-        )
+        return ((self.ixmin == other.ixmin)
+                and (self.ixmax == other.ixmax)
+                and (self.iymin == other.iymin)
+                and (self.iymax == other.iymax))
 
     def __or__(self, other):
         return self.union(other)
@@ -147,14 +144,22 @@ class BoundingBox:
         return fmt.format(**data)
 
     @property
+    def center(self):
+        """
+        The ``(y, x)`` center of the bounding box.
+        """
+        return (0.5 * (self.iymax - 1 + self.iymin),
+                0.5 * (self.ixmax - 1 + self.ixmin))
+
+    @property
     def shape(self):
         """
         The ``(ny, nx)`` shape of the bounding box.
         """
-
         return self.iymax - self.iymin, self.ixmax - self.ixmin
 
     @property
+    @deprecated('1.1', alternative='get_overlap_slices')
     def slices(self):
         """
         The bounding box as a tuple of `slice` objects.
@@ -162,8 +167,56 @@ class BoundingBox:
         The slice tuple is in numpy axis order (i.e., ``(y, x)``) and
         therefore can be used to slice numpy arrays.
         """
-
+        if self.iymin < 0 or self.ixmin < 0:
+            return ValueError('cannot create slices when ixmin or iymin '
+                              'is negative')
         return (slice(self.iymin, self.iymax), slice(self.ixmin, self.ixmax))
+
+    def get_overlap_slices(self, shape):
+        """
+        Get slices for the overlapping part of the bounding box and an
+        2D array.
+
+        Parameters
+        ----------
+        shape : 2-tuple of int
+            The shape of the 2D array.
+
+        Returns
+        -------
+        slices_large : tuple of slices or `None`
+            A tuple of slice objects for each axis of the large array,
+            such that ``large_array[slices_large]`` extracts the region
+            of the large array that overlaps with the small array.
+            `None` is returned if there is no overlap of the bounding
+            box with the given image shape.
+        slices_small : tuple of slices or `None`
+            A tuple of slice objects for each axis of an array enclosed
+            by the bounding box such that ``small_array[slices_small]``
+            extracts the region that is inside the large array. `None`
+            is returned if there is no overlap of the bounding box with
+            the given image shape.
+        """
+        if len(shape) != 2:
+            raise ValueError('input shape must have 2 elements.')
+
+        xmin = self.ixmin
+        xmax = self.ixmax
+        ymin = self.iymin
+        ymax = self.iymax
+
+        if xmin >= shape[1] or ymin >= shape[0] or xmax <= 0 or ymax <= 0:
+            # no overlap of the bounding box with the input shape
+            return None, None
+
+        slices_large = (slice(max(ymin, 0), min(ymax, shape[0])),
+                        slice(max(xmin, 0), min(xmax, shape[1])))
+        slices_small = (slice(max(-ymin, 0),
+                              min(ymax - ymin, shape[0] - ymin)),
+                        slice(max(-xmin, 0),
+                              min(xmax - xmin, shape[1] - xmin)))
+
+        return slices_large, slices_small
 
     @property
     def extent(self):
@@ -178,13 +231,8 @@ class BoundingBox:
         indexing. This is useful for plotting the bounding box using
         Matplotlib.
         """
-
-        return (
-            self.ixmin - 0.5,
-            self.ixmax - 0.5,
-            self.iymin - 0.5,
-            self.iymax - 0.5,
-        )
+        return (self.ixmin - 0.5, self.ixmax - 0.5,
+                self.iymin - 0.5, self.iymax - 0.5)
 
     @deprecated('0.7', alternative='as_artist')
     def as_patch(self, **kwargs):  # pragma: no cover
@@ -203,7 +251,6 @@ class BoundingBox:
         result : `matplotlib.patches.Rectangle`
             A matplotlib rectangular patch.
         """
-
         return self.as_artist(**kwargs)
 
     def as_artist(self, **kwargs):
@@ -238,7 +285,6 @@ class BoundingBox:
             ax.add_patch(bbox.as_artist(facecolor='none', edgecolor='white',
                          lw=2.))
         """
-
         from matplotlib.patches import Rectangle
 
         return Rectangle(xy=(self.extent[0], self.extent[2]),
@@ -249,14 +295,10 @@ class BoundingBox:
         Return a `~photutils.aperture.RectangularAperture` that
         represents the bounding box.
         """
-
         from .rectangle import RectangularAperture  # prevent circular import
 
-        xpos = (self.extent[1] + self.extent[0]) / 2.
-        ypos = (self.extent[3] + self.extent[2]) / 2.
-        xypos = (xpos, ypos)
+        xypos = self.center[::-1]  # xy order
         height, width = self.shape
-
         return RectangularAperture(xypos, w=width, h=height, theta=0.)
 
     def plot(self, axes=None, origin=(0, 0), **kwargs):
@@ -277,7 +319,6 @@ class BoundingBox:
         kwargs : `dict`
             Any keyword arguments accepted by `matplotlib.patches.Patch`.
         """
-
         aper = self.to_aperture()
         aper.plot(axes=axes, origin=origin, **kwargs)
 
@@ -297,7 +338,6 @@ class BoundingBox:
             A `BoundingBox` representing the union of the input
             `BoundingBox` with this one.
         """
-
         if not isinstance(other, BoundingBox):
             raise TypeError('BoundingBox can be joined only with another '
                             'BoundingBox.')
@@ -325,7 +365,6 @@ class BoundingBox:
             A `BoundingBox` representing the intersection of the input
             `BoundingBox` with this one.
         """
-
         if not isinstance(other, BoundingBox):
             raise TypeError('BoundingBox can be intersected only with '
                             'another BoundingBox.')